adjust tests to renamed files, remove hardcoded lines
